I feel that optionsForSize belongs toget with getIconSvg
