+++ lib/generateIcon.js
@@ -5,34 +5,17 @@ var Promise = require("bluebird"),
   pathModule = require('path'),
   spawn = require('child_process').spawn,
   execFile = Promise.promisify(require('child_process').execFile),
-  getTemplate = require("./getTemplate");
-
+  getIconSvg = require("./getIconSvg");
 
 var svgo = new SVGO({
   removeViewBox: true
 });
 
-var PIXEL = 128;
-function optionsForSize(siz, addPadding) {
-  var padding = 0;
-  if(addPadding) {
-    var pad = parseInt(siz / 14)*14;
-    padding = ((siz - pad) / 2) * (PIXEL/2);
-  }
-
-  return {
-    paddingTop: padding,
-    paddingBottom: padding,
-    paddingLeft: padding,
-    paddingRight: padding
-  };
-}
-
 function generatePng(siz, name, params) {
   return new Promise(function(resolve, reject) {
     var rsvgConvert;
     var color = params.color;
-    var svgCode = getTemplate(optionsForSize(siz, params.addPadding), params);
+    var svgCode = getIconSvg(params, siz);
     var filename = pathModule.join(params.destFolder, color, 'png', siz.toString(), name + '.png');
     rsvgConvert = spawn('rsvg-convert', ['-f', 'png', '-w', siz, '-o', filename]);
     if (process.env.INTERMEDIATE_SVG) {
@@ -67,12 +50,7 @@ function generateSvg(name, params) {
 
   return new Promise(function(resolve, reject) {
     var outSvg = fs.createWriteStream(pathModule.join(svgFolder, name + '.svg'));
-    svgo.optimize(getTemplate({
-      paddingTop: 0,
-      paddingLeft: 0,
-      paddingBottom: 0,
-      paddingRight: 0
-    }, params), function(result) {
+    svgo.optimize(getIconSvg(params), function(result) {
       outSvg.end(result.data);
       resolve();
     });
