+++ lib/generateSprites.js
@@ -3,7 +3,7 @@ var Promise = require("bluebird"),
   SVGO = require('svgo'),
   pathModule = require('path'),
   mkdirp = require("mkdirp"),
-  getTemplate = require("./getTemplate");
+  getIconSvg = require("./getIconSvg");
 
 var svgo = new SVGO({
   removeViewBox: true
@@ -12,12 +12,7 @@ var svgo = new SVGO({
 function generateSprite(name, params) {
 
   return new Promise(function(resolve, reject) {
-    svgo.optimize(getTemplate({
-      paddingTop: 0,
-      paddingLeft: 0,
-      paddingBottom: 0,
-      paddingRight: 0
-    }, params), function(result) {
+    svgo.optimize(getIconSvg(params), function(result) {
 
       var m = result.data.match('(<path.*\/>)');
