+++ lib/getIconSvg.js
@@ -9,8 +9,25 @@ var template =
   '</svg>';
 
 var PIXEL = 128;
+function optionsForSize(siz, addPadding) {
+  var padding = 0;
 
-function getTemplate(options, params) {
+  if(addPadding && siz) {
+    var pad = parseInt(siz / 14)*14;
+    padding = ((siz - pad) / 2) * (PIXEL/2);
+  }
+
+  return {
+    paddingTop: padding,
+    paddingBottom: padding,
+    paddingLeft: padding,
+    paddingRight: padding
+  };
+}
+function getIconSvg(params, size) {
+
+  var options = optionsForSize(size, params.addPadding);
+
   var out = template.substr(0);
   params = extend({}, params, {
     shiftX: -(-(14*PIXEL - params.advWidth)/2 - options.paddingLeft),
@@ -25,4 +42,4 @@ function getTemplate(options, params) {
   return out;
 }
 
-module.exports = getTemplate;
+module.exports = getIconSvg;
